# Conflicts:
#	.idea/compiler.xml
#	.idea/misc.xml
#	backend-hvz/src/main/java/com/example/backendhvz/controllers/GameController.java
#	backend-hvz/src/main/java/com/example/backendhvz/services/game/GameService.java
#	backend-hvz/src/main/java/com/example/backendhvz/services/game/GameServiceImpl.java

+++ backend-hvz/src/main/java/com/example/backendhvz/services/game/GameServiceImpl.java
@@ -1,7 +1,8 @@
 package com.example.backendhvz.services.game;
 
-import com.example.backendhvz.enums.GameState;
 import com.example.backendhvz.enums.PlayerState;
+import com.example.backendhvz.exceptions.ForbiddenException;
+import com.example.backendhvz.exceptions.NotFoundException;
 import com.example.backendhvz.models.Game;
 import com.example.backendhvz.models.Player;
 import com.example.backendhvz.repositories.GameRepository;
@@ -23,6 +24,7 @@ public class GameServiceImpl implements GameService {
 
     @Override
     public Game findById(Long gameId) {
+        if(!gameRepository.existsById(gameId)) throw new NotFoundException("Game with ID " + gameId + " not found.");
         return gameRepository.findById(gameId).get();
     }
 
@@ -52,32 +54,39 @@ public class GameServiceImpl implements GameService {
 
     @Override
     public Game update(Game game) {
+        if(!gameRepository.existsById(game.getId())) throw new NotFoundException("Game with ID " + game.getId() + " not found.");
         return gameRepository.save(game);
     }
 
     @Override
-    public Game updateGame(Game game, Long updatingPlayerId) {
+    public Game updateGame(Long updatingPlayerId, Game game) {
+        if(!gameRepository.existsById(game.getId())) throw new NotFoundException("Game with ID " + game.getId() + " not found.");
+        if(!playerRepository.existsById(updatingPlayerId)) throw new NotFoundException("Player with ID " + updatingPlayerId + " not found.");
+
         Player updatingPlayer = playerRepository.findById(updatingPlayerId).get();
-        if(updatingPlayer.getState() != PlayerState.ADMINISTRATOR) return null;
+        if(updatingPlayer.getState() != PlayerState.ADMINISTRATOR) throw new ForbiddenException("Only administrators can update games.");
         return gameRepository.save(game);
     }
 
     @Override
     public void deleteById(Long gameId) {
+        if(!gameRepository.existsById(gameId)) throw new NotFoundException("Game with ID " + gameId + " not found.");
         gameRepository.deleteById(gameId);
     }
 
     @Override
     public void deleteGameById(Long gameId, Long deletingPlayerId) {
+        if(!gameRepository.existsById(gameId)) throw new NotFoundException("Game with ID " + gameId + " not found.");
+        if(!playerRepository.existsById(deletingPlayerId)) throw new NotFoundException("Player with ID " + deletingPlayerId + " not found.");
+
         Player deletingPlayer = playerRepository.findById(deletingPlayerId).get();
-        if(deletingPlayer.getState() != PlayerState.ADMINISTRATOR) return;
+        if(deletingPlayer.getState() != PlayerState.ADMINISTRATOR) throw new ForbiddenException("Only administratos can delete games.");
         gameRepository.deleteById(gameId);
     }
 
-
-
     @Override
     public void delete(Game game) {
+        if(!gameRepository.existsById(game.getId())) throw new NotFoundException("Game with ID " + game.getId() + " not found.");
         gameRepository.delete(game);
     }
 }
